Rename changelog_release -> changelog_release_utils; trim CHANGELOG header
Module rename matches the existing `utils.py` convention: helpers that
sit alongside top-level scripts in `scripts/` carry a `_utils` suffix.
Also drops the meta paragraph from CHANGELOG.md's header that described
how the cron and release.py manage [Unreleased] -- that's implementation
detail readers don't need to see.
