+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-A concise, human-friendly summary of changes. The nightly changelog consolidation agent appends bullets to `[Unreleased]`, and `scripts/release.py` finalizes that section into a versioned heading at release time. Entries are categorized using the [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) categories: Added, Changed, Deprecated, Removed, Fixed, Security.
+A concise, human-friendly summary of changes. Entries are categorized using the [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) categories: Added, Changed, Deprecated, Removed, Fixed, Security.
 
 For the full, unedited changelog entries, see [UNABRIDGED_CHANGELOG.md](UNABRIDGED_CHANGELOG.md).
